fix(mobile): excluir dev-client pelo package.json — o exclude do gradle é quebrado
A inspeção do APK de preview 0.30.1 mostrou que o fix anterior não funcionou: 260
referências a expo/modules/devlauncher no classes2.dex, a Activity do ML Kit ainda
declarada no manifest e os modelos .tflite ainda empacotados.

A causa é upstream, em AutolinkigCommandBuilder.kt:44-46. O `option(key, List)`
achata a lista com joinToString(" ") e entrega tudo como um argumento só, mas a
flag do CLI é variádica (--exclude <exclude...>). O autolinking acaba procurando
um módulo literalmente chamado "expo-dev-client expo-dev-launcher expo-dev-menu
expo-dev-menu-interface", não encontra, e exclui zero — sem erro e sem aviso.

Reproduzido lado a lado:

  --exclude "a b c d"        → 27 módulos, 4 dev   (o que o gradle faz)
  --exclude a --exclude b …  → 23 módulos, 0 dev   (o que eu tinha testado)

Validei o mecanismo pela CLI e presumi que o build o invocaria do mesmo jeito. O
settings.gradle estava certo, a CLI estava certa, e o produto estava errado. Quem
pegou foi a inspeção do artefato — a própria R-308 escrita dois commits atrás.

Também medido: excluir um módulo não derruba os outros, porque são resolvidos
independentemente; e mover o pacote para devDependencies não muda nada.

O fix agora usa expo.autolinking.android.exclude no package.json, que o CLI lê
direto e não passa pelo builder quebrado. A prova passa a ser `gradlew projects`,
não a CLI: 23 → 19 projetos, nenhum expo-dev-*. O iOS fica intocado.

Custo aceito: o exclude é estático, porque não existe variável de ambiente que
condicione autolinking por perfil. O dev build Android perde o menu do Expo; o do
React Native continua, e nada no repositório usava --dev-client.

O plugin fica só com o que de fato funcionou e continua necessário — remover o
scheme exp+<slug> do manifest de release, que vem do config plugin do dev-client
e por isso sobrevive à exclusão dos módulos. Renomeado para refletir o escopo.

AP-324 e R-308 corrigidos: descreviam um fix que não funcionava. R-308 ganhou a
distinção que faltava entre ferramenta de consulta e ferramenta de build.
