+++ .agent/memory/RULES_INDEX.md
@@ -2,7 +2,7 @@
 
 ## 🚀 Release & Versionamento
 
-- **[R-308]** 🔴 **O conteúdo do binário de release se INSPECIONA, não se presume — e o build de smoke tem que ser o mesmo binário.** Nenhum gate do repo enxerga o que entra no APK/AAB: `tsc`/`eslint`/jest/`strict-island.sh` operam sobre JS/TS e não veem o grafo do gradle nem o manifest merged; `expo-doctor` valida versão, não conteúdo. Logo toda mudança de composição nativa (dependência, autolinking, minificação, manifest, permissão, plugin de build) só está verificada quando o ARTEFATO foi lido — escada barata→cara: `expo-modules-autolinking resolve -p android --json` → ler `settings.gradle`/`gradle.properties`/`AndroidManifest.xml` pós-prebuild → `aapt2 dump badging` → avisos do Play Console. §2: o perfil usado no SMOKE recebe a MESMA config de release que o de produção — se `preview` e `production` divergem (minify, exclusão de módulo, permissão), o smoke valida um artefato que não vai à loja, e ele era a única verificação existente (R8 quebra reflexão só em runtime: APK não-minificado passa feliz e o AAB minificado quebra em produção). Divergência entre perfis se declara, não se herda. Complementa [[R-307]]: R-307 responde *de qual código saiu*, R-308 *o que tem dentro*. Origem: 3 avisos do Play sobre o AAB da v0.30.0 que passaram por todos os gates ([[AP-324]]) -> [`rules/mobile_and_platform/R-308.md`](./rules/mobile_and_platform/R-308.md)
+- **[R-308]** 🔴 **O conteúdo do binário de release se INSPECIONA, não se presume — e o build de smoke tem que ser o mesmo binário.** Nenhum gate do repo enxerga o que entra no APK/AAB: `tsc`/`eslint`/jest/`strict-island.sh` operam sobre JS/TS e não veem o grafo do gradle nem o manifest merged; `expo-doctor` valida versão, não conteúdo. Logo toda mudança de composição nativa (dependência, autolinking, minificação, manifest, permissão, plugin de build) só está verificada quando o ARTEFATO foi lido — escada barata→cara: `cd android && ./gradlew projects` (o que o BUILD inclui) → ler `settings.gradle`/`gradle.properties`/`AndroidManifest.xml` pós-prebuild → `aapt2 dump badging` + `strings` nos `classes*.dex` → avisos do Play Console. **⚠️ Ferramenta de consulta ≠ ferramenta de build:** `expo-modules-autolinking resolve` diz o que o CLI resolveria, NÃO o que o gradle compila — os dois divergem quando a ponte tem bug, e foi assim que o fix do [[AP-324]] passou por todos os gates e saiu quebrado no APK (settings.gradle certo, CLI certo, produto errado). Validar por proxy é o modo de falha DESTA regra. §2: o perfil usado no SMOKE recebe a MESMA config de release que o de produção — se `preview` e `production` divergem (minify, exclusão de módulo, permissão), o smoke valida um artefato que não vai à loja, e ele era a única verificação existente (R8 quebra reflexão só em runtime: APK não-minificado passa feliz e o AAB minificado quebra em produção). Divergência entre perfis se declara, não se herda. Complementa [[R-307]]: R-307 responde *de qual código saiu*, R-308 *o que tem dentro*. Origem: 3 avisos do Play sobre o AAB da v0.30.0 que passaram por todos os gates ([[AP-324]]) -> [`rules/mobile_and_platform/R-308.md`](./rules/mobile_and_platform/R-308.md)
 - **[R-307]** 🔴 **Binário publicado tem procedência rastreável — a versão vive no GIT, não só no `app.config.js`.** `APP_VERSION` diz QUAL versão saiu; não diz QUAL CÓDIGO virou aquela versão, e três capacidades dependem desse elo: reproduzir um build antigo para investigar bug, publicar hotfix por OTA, e auditar o que roda em campo. Build de loja exige **árvore limpa** (binário de árvore suja é irrastreável — e não se corrige por OTA) e recebe **tag `mobile-v<APP_VERSION>` no commit que o gerou**, criada no momento do build (depois é adivinhação) e **empurrada pro origin** (tag local some com a máquina). Uma tag por versão, não por plataforma — colisão em commit diferente é erro duro: código mudou sem bumpar ([[R-221]] §4), bumpe, **nunca mova a tag**. **§3b (emenda 2026-07-30):** a tag marca o commit COMPILADO; commits posteriores ficam fora **por definição, inclusive os que não afetam o binário** (docs/scripts/memória) — não existe exceção p/ "delta fora do bundle", porque a tag responde *qual código virou este binário* e um commit feito depois do build não virou; mover faz a tag apontar p/ commit nunca compilado. É NORMAL a tag ficar atrás do HEAD da main — não é dívida. **OTA de produção só sai dessa tag ou de `hotfix/ota-*` cortado dela**: o `eas update` empacota a WORKING TREE, então publicar da `main` que já andou entrega features não lançadas junto do fix — código não revisado, possivelmente chamando nativo que o binário não tem (crash), e violação da Apple 3.3.1. Reproduzir build antigo = `checkout` da tag + `npm ci` + `build-*.sh` (funciona porque o lock é versionado e o nativo é regenerado pelo prebuild). ⚠️ Limite honesto: a tag reproduz o CÓDIGO, não a toolchain (Xcode/SDK não são fixados pelo repo). Gates: `lib-release-tag.sh` nos 2 scripts de build + guarda em `publish-ota.sh`. Origem: 051-A PR 1.6 — 28 tags no repo, todas da web, nenhuma de build mobile -> [`rules/mobile_and_platform/R-307.md`](./rules/mobile_and_platform/R-307.md)
 
 ## 📦 Data & Schema (`data_and_schema`)
