+++ .agent/memory/anti-patterns/mobile_and_platform/AP-324.md
@@ -65,32 +65,78 @@ gate** neste repo. Toda dependência de desenvolvimento é candidata ao mesmo va
 - Aviso da loja apontando classe de um pacote que você não chama.
 - Permissão, Activity ou scheme no manifest de produção que ninguém no time reconhece.
 
-## Fix aplicado (2026-07-30)
+## 🔴 A primeira tentativa de fix FALHOU — e o motivo é a parte mais útil deste AP
 
-Config plugin `apps/mobile/withoutDevClientOnRelease.js`: usa `expoAutolinking.exclude`
-(propriedade da extensão do settings plugin, vira `--exclude` no comando de autolinking) para tirar
-`expo-dev-client`/`-launcher`/`-menu`/`-menu-interface` quando `EAS_BUILD_PROFILE` está definido e é
-diferente de `development`. Remove também o scheme `exp+<slug>`, que vem do config plugin do
-dev-client e portanto sobrevive à exclusão do autolinking.
+A tentativa inicial usou `expoAutolinking.exclude` no `settings.gradle`, escrito por um config
+plugin. O `settings.gradle` gerado continha a linha correta, e o CLI de autolinking respeitava
+`--exclude`. **O APK saiu com o dev-launcher dentro assim mesmo** — 260 referências a
+`expo/modules/devlauncher` no `classes2.dex`, `GmsBarcodeScanningDelegateActivity` ainda no
+manifest, modelos `.tflite` do ML Kit ainda empacotados.
 
-Medido: **27 módulos → 23** resolvidos para Android.
+**Causa: bug upstream no plugin gradle do Expo.** Em `AutolinkigCommandBuilder.kt:44-46`:
 
-**`preview` entra junto de propósito.** Se a exclusão valesse só em `production`, o APK usado no
-smoke deixaria de reproduzir o binário publicado — e, como nada disto é verificável por teste, o
-smoke é a única verificação que existe. Um smoke que valida um artefato diferente do que vai à loja
-não é smoke, é cerimônia.
-
-## Verificação
-
-```bash
-cd apps/mobile
-npx expo-modules-autolinking resolve -p android --json | \
-  node -e "let s='';process.stdin.on('data',d=>s+=d).on('end',()=>{const m=JSON.parse(s).modules.map(x=>x.packageName||x.name);console.log(m.length, m.filter(n=>/expo-dev-/.test(n)))})"
+```kotlin
+fun option(key: String, value: List<String>) = apply {
+  optionsMap[key] = value.joinToString(" ")   // ← lista vira UMA string
+}
 ```
 
-Sem `EAS_BUILD_PROFILE` deve listar os 4 módulos de dev (é o dev build). Com
-`EAS_BUILD_PROFILE=production`, o `settings.gradle` pós-prebuild deve conter a linha
-`expoAutolinking.exclude = [...]`.
+A flag do CLI é variádica (`--exclude <exclude...>`). Recebendo um argumento único, o autolinking
+procura por um módulo literalmente chamado
+`"expo-dev-client expo-dev-launcher expo-dev-menu expo-dev-menu-interface"` — que não existe.
+Exclui zero, sem erro, sem aviso.
+
+Reprodução lado a lado:
+
+```
+--exclude "a b c d"         → 27 módulos, 4 dev   ← o que o gradle faz
+--exclude a --exclude b …   → 23 módulos, 0 dev   ← o que foi testado
+```
+
+**A lição:** validar um mecanismo pela CLI e presumir que o build o invoca do mesmo jeito. O
+`settings.gradle` estava certo, o CLI estava certo, e o produto estava errado — porque a ponte
+entre os dois não era o que se supunha. Só a inspeção do artefato pegou ([[R-308]]).
+
+Excluir um módulo só também não resolve: são resolvidos independentemente
+(`--exclude expo-dev-client` → 26 módulos, 3 dev restantes).
+
+## Fix que funciona (2026-07-30)
+
+**1. Exclusão dos módulos — `apps/mobile/package.json`:**
+
+```json
+"expo": { "autolinking": { "android": { "exclude": [
+  "expo-dev-client", "expo-dev-launcher", "expo-dev-menu", "expo-dev-menu-interface"
+] } } }
+```
+
+O CLI lê o `package.json` diretamente (`parsePackageJsonOptions`), sem passar pelo builder
+quebrado. Escopo `android` deliberado: o iOS não é afetado (28 módulos antes e depois) e lá o
+`debugOnly` já resolve.
+
+**Custo aceito pelo PO:** é estático — não há env var que condicione autolinking por perfil, então
+vale também no dev build Android. `expo run:android` perde o dev launcher e o dev menu do Expo; o
+dev menu do próprio React Native (reload, inspector) permanece. Nenhum script do repo usa
+`--dev-client`, e o perfil `development` do `eas.json` não seta `developmentClient`.
+
+**2. Scheme de dev — `apps/mobile/withoutDevSchemeOnRelease.js`:** o `exp+<slug>` NÃO vem do
+autolinking (é do config plugin do dev-client) e por isso sobrevive à exclusão dos módulos. O
+plugin o remove do manifest de release, abortando se o scheme `dosiq` não sobreviver.
+
+## Verificação — use o gradle, não o CLI
+
+A verificação pelo CLI **passa mesmo com o bug** e foi exatamente o que produziu o falso positivo.
+A pergunta certa é o que o build monta:
+
+```bash
+cd apps/mobile/android && ./gradlew projects | grep -c "expo-dev-"   # tem que ser 0
+```
+
+Medido após o fix: 23 → **19 projetos**, nenhum `expo-dev-*`. Confirmação final no artefato:
+
+```bash
+unzip -p <apk> classes2.dex | strings | grep -c "expo/modules/devlauncher"   # tem que ser 0
+```
 
 ## Relacionados
