+++ .agent/memory/rules/mobile_and_platform/R-308.md
@@ -33,10 +33,17 @@ Formas aceitáveis de inspeção, da mais barata à mais cara:
 
 | Pergunta | Como responder |
 |---|---|
-| que módulos nativos entram? | `npx expo-modules-autolinking resolve -p android --json` |
+| que módulos nativos o BUILD inclui? | `cd android && ./gradlew projects` |
 | o que o prebuild gerou? | ler `android/settings.gradle`, `android/gradle.properties`, `AndroidManifest.xml` |
-| o que foi realmente empacotado? | `aapt2 dump badging <arquivo>` / abrir o AAB |
+| o que foi realmente empacotado? | `aapt2 dump badging` · `unzip -l` · `strings` nos `classes*.dex` |
 | o que a loja vê? | avisos do Play Console após upload no internal testing |
+
+**⚠️ Ferramenta de consulta ≠ ferramenta de build.** `npx expo-modules-autolinking resolve` responde
+*"o que o CLI resolveria"*, **não** *"o que o gradle vai compilar"* — os dois divergem quando a
+ponte entre eles tem bug. Foi exatamente assim que o fix do [[AP-324]] passou por todos os gates e
+saiu quebrado no APK: `settings.gradle` correto, CLI correto, produto errado. **Validar por
+proxy é o modo de falha desta regra**; a pergunta "que módulos entram?" só é respondida por
+`./gradlew projects`, e a resposta final está sempre no artefato.
 
 **2. O perfil de build usado no smoke recebe a MESMA configuração de release que o de produção.**
 Se `preview` e `production` divergem em minificação, exclusão de módulos ou permissões, o smoke
